requirements.txt
flask==3.1.0
tensorflow==2.17.1
gunicorn==21.2.0
python-dotenv==1.0.1